move  verifyTreasureBoxRequest as util func
