+++ src/app/api/treasure-box/route.ts
@@ -1,3 +1,4 @@
+import { verifyTreasureBoxRequest } from '@/utils/verifyTreasureBoxRequest';
 import { PrismaClient } from '@prisma/client';
 import { NextResponse, type NextRequest } from 'next/server';
 
@@ -78,25 +79,4 @@ export async function POST(request: NextRequest) {
   }
 
   return NextResponse.json({ status: 'ok' });
-}
-
-export function verifyTreasureBoxRequest(body: TreasureBoxType) {
-  const { gameId, points, user } = body;
-
-  // check if it has all the required fields
-  if (!gameId) {
-    return new Response('Missing parameters: gameId', { status: 400 });
-  }
-
-  if (!points) {
-    return new Response('Missing parameters: points', { status: 400 });
-  }
-
-  if (!user) {
-    return new Response('Missing parameters: user', { status: 400 });
-  }
-
-  if (!user.address) {
-    return new Response('Missing parameters: user address', { status: 400 });
-  }
 }
