+++ src/hooks/useMutateTreasureBox.tsx
@@ -1,10 +1,8 @@
 import { useMutation } from 'react-query';
-import {
-  TreasureBoxType,
-  verifyTreasureBoxRequest,
-} from '@/app/api/treasure-box/route';
+import { TreasureBoxType } from '@/app/api/treasure-box/route';
 import { routes } from '@/constants/routes';
 import axios from 'axios';
+import { verifyTreasureBoxRequest } from '@/utils/verifyTreasureBoxRequest';
 
 export function useMutateTreasureBox() {
   const attackBox = useMutation({
